% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greedy_knapsack.R
\name{greedy_knapsack}
\alias{greedy_knapsack}
\title{Greedy heuristic algorithm Knapsack Solver}
\usage{
greedy_knapsack(x, W)
}
\arguments{
\item{x}{A data frame with two columns: \code{w} (weights) and \code{v} (values) of the items.}

\item{W}{The total capacity of the knapsack.}
}
\value{
A list containing:
\item{value}{The total value of the selected items, rounded to the nearest integer.}
\item{elements}{A vector of the original indices of the selected items.}
}
\description{
Greedy heuristic algorithm Knapsack Solver
}
\details{
The algorithm follows a greedy heuristic, which sorts the items
by their value-to-weight ratio in descending order and adds them to the
knapsack if their inclusion does not exceed the weight capacity.
This approach does not guarantee an optimal solution but is efficient.
}
\references{
https://en.wikipedia.org/wiki/Knapsack_problem#Greedy_approximation_algorithm
}
